Rename RatingBase.tsx to Rating.base.tsx to match convention
