+++ packages/office-ui-fabric-react/src/components/Rating/Rating.tsx
@@ -3,7 +3,7 @@ import {
 } from '../../Utilities';
 import { IRatingProps } from './Rating.types';
 import { getStyles } from './Rating.styles';
-import { RatingBase } from './RatingBase';
+import { RatingBase } from './Rating.base';
 
 export const Rating = styled(
   RatingBase,
